dao와 dto 수정 및 업데이트 구직자 이력서 등록 초안 완성

+++ rest/src/main/java/com/jamjam/rest/dto/Role.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Role {
 
-	private Long role_id;
-	private Long user_id;
+	private Integer role_id;
+	private Integer user_id;
 	private Integer company_id;
 	private String role;
 }
